Move community structure classes into their own header
